Split renderer.js into various sub files
