+++ src/renderer/state1.js
@@ -1,0 +1,9 @@
+function setup() {
+	loadMenu();
+	render();
+}
+
+function loadMenu() {
+	menu = new PIXI.Sprite(PIXI.loader.resources.menu.texture);
+	stage.addChild(menu);
+}
